refactor out two tests
